+++ google-maps/e2e-tests/src/pages/Scrolling/SimpleScrolling.tsx
@@ -122,7 +122,7 @@ const SimpleScrollingPage: React.FC = () => {
               height: 75,
               backgroundColor: 'blue',
               position: 'relative',
-              top: 0,
+              top: -10,
               left: 40,
               zIndex: 1000,
               opacity: 0.5,